% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coExp_network.R
\name{coExp_network}
\alias{coExp_network}
\title{Construct coexpression network and sub-community detection.}
\usage{
coExp_network(dat1, min_size=2, n_network=2)
}
\arguments{
\item{edge_dat}{An edge list dataframe. Recommend to use the RS score generate from robustness function but can be apply to any correlation measurements.}

\item{min_size}{Minimum size of networks to include in the communty detection.}

\item{n_networks}{Number of sub-communities to visualise.}
}
\description{
Construct coexpression network and sub-community detection.
}
\author{
Huiwen Zheng
}
