% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoDis.R
\name{plotCoDis}
\alias{plotCoDis}
\title{Visualisation of selected coexpressed gene pairs' correlation distribution}
\usage{
plotCoDis(dat_list, gene_name1, gene_name2, group_name = "sample")
}
\arguments{
\item{gene_name1}{Gene1 to plot}

\item{gene_name2}{Gene2 to plot}

\item{prop_list}{A list of prop_distribution dataframes.}
}
\value{
A ggplot object with violin plot
}
\description{
Visualisation of selected coexpressed gene pairs' correlation distribution
}
\author{
Huiwen Zheng
}
