% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoExp.R
\name{plotCoExp}
\alias{plotCoExp}
\title{Visualise correlated gene pairs with transformed data}
\usage{
plotCoExp(dat, gene1, gene2)
}
\arguments{
\item{dat}{Transformed data, encourage to use the transformed data from prop_distribution function but can be applied to any data.}

\item{gene1}{A gene to plot}

\item{gene2}{A gene to plot}
}
\description{
Visualise correlated gene pairs with transformed data
}
\author{
Huiwen Zheng
}
