% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPways.R
\name{plotPways}
\alias{plotPways}
\title{Visualise pathway enrichment results}
\usage{
plotPways(pways_df, n = 10)
}
\arguments{
\item{pways_df}{Output from pathway analysis.}

\item{n}{Number of pathways plotted in the barplot.}
}
\description{
Visualise pathway enrichment results
}
\author{
Huiwen Zheng
}
