% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.R
\name{robustness}
\alias{robustness}
\title{Compute statstical correlation robustness test}
\usage{
robustness(prop_list, stats = T, p.adj = 0.1)
}
\arguments{
\item{prop_list}{A list generated by prop_distribution function, which at least contains bi_zero and prop_distribution.}

\item{stats}{By default, performs asymptotic test for the equality of coefficients of variation to determine against null distribution.          If changes to FALSE, top *1%* most stably correlated gene pairs will be returned.}

\item{p.adj}{Significance level used in the asymptotic test.}
}
\value{
A dataframe with gene pair information, robustness score *(RS)* and p.adj value.
}
\description{
Compute statstical correlation robustness test
}
\author{
Huiwen Zheng
}
