% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness2.R
\name{robustness2}
\alias{robustness2}
\title{Compute statistical correlation robustness test between two samples}
\usage{
robustness2(group_list, ref = 1, p.adj = 0.05)
}
\arguments{
\item{group_list}{A list that contains multiple prop_distribution list generated by prop_distribution function, which at least contains bi_zero and prop_distribution.}

\item{ref}{An integer indicates the reference used in the comparison. Default set to 1, indicating the first proportion list is used as reference.}

\item{p.adj}{Significance level used in the asymptotic test.}
}
\value{
A dataframe with gene pair names, type (either changed in sign or through statistical testing), absolute fold-change.
}
\description{
Compute statistical correlation robustness test between two samples
}
\author{
Huiwen Zheng
}
