% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness2.R
\name{robustness2}
\alias{robustness2}
\title{Compute statistical correlation robustness test between two samples}
\usage{
robustness2(prop_list1, prop_list2, p.adj = 0.05)
}
\arguments{
\item{prop_list1}{A list generated by prop_distribution function, which at least contains bi_zero and prop_distribution.
#' @param prop_list2 A list generated by prop_distribution function, which at least contains bi_zero and prop_distribution.}

\item{p.adj}{Significance level used in the asymptotic test.}
}
\value{
A dataframe with gene pair information, robustness score *(RS)*
}
\description{
Compute statistical correlation robustness test between two samples
}
\author{
Huiwen Zheng
}
